task 15.1

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -8,7 +8,7 @@ int main(void)
 	unsigned long a = 1;
 	unsigned long b = 2;
 	unsigned long c;
-	int sum = 2;
+	unsigned long sum = 2;
 
 	printf("%ld, %ld, ", a, b);
 	while (1)
